refactor(schema): move candidateTriggersFor para lib/conditional (react-doctor)
only-export-components: ConditionEditor.tsx exportava a funcao pura
candidateTriggersFor junto do componente, quebrando fast-refresh. Movida
para lib/conditional.ts (modulo puro, ja lida com condicoes/campos);
FieldCard e EditFieldDialog importam de la.

Os outros 3 only-export-components (ui/button, ui/badge, ui/tabs
exportando *Variants) sao convencao shadcn/ui e ficam como estao.
